requirements.txt
langchain==0.2.0
langchain-community==0.2.0
langchain-google-genai==2.1.5
langchain-huggingface==0.3.0
langgraph==0.2.0
streamlit==1.29.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
google-generativeai==0.3.2 